+++ passivetotal/analyzer/hostname.py
@@ -14,14 +14,14 @@ from passivetotal.analyzer.trackers import HasTrackers
 from passivetotal.analyzer.components import HasComponents
 from passivetotal.analyzer.illuminate import HasReputation
 from passivetotal.analyzer.articles import HasArticles
-from passivetotal.analyzer.enrich import HasMalware
+from passivetotal.analyzer.enrich import HasMalware, HasSubdomains
 from passivetotal.analyzer.projects import IsArtifact
 
 
 
 class Hostname(HasComponents, HasCookies, HasTrackers, HasHostpairs, 
                HasReputation, HasArticles, HasResolutions, HasSummary,
-               HasMalware, IsArtifact):
+               HasMalware, HasSubdomains, IsArtifact):
 
     """Represents a hostname such as api.passivetotal.org.